Moved the Update enum to program/update.rs
